+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2024-02-27 - v2.0.0
+### Breaking changes
+- Switch to [ESM module](https://gist.github.com/sindresorhus/a39789f98801d908bbc7ff3ecc99d99c)
+
+### Changes
+- Now possible to add options for fetch (node-fetch) requests via (MyContentClient).fetchRequestHeaders and (MyContentClient).fetchOptions
+- Updated packages to latest major versions
+
 # 2023-04-03 - v1.1.0
 ### changes
 - Added schedule json to post and edit